backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
alembic>=1.13.1
python-multipart>=0.0.9
pandas>=2.2.2
openpyxl>=3.1.2
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
email-validator>=2.1.1
